improvement
1. xmodmap : remapped Caps Lock. it nows acts as control and
shift+caps_lock acts as caps lock

2. took a screenshot of my new config

3. removed outdated/ as I'd unvoluntarily duplicated it. usbmounter.sh
can still be found in devices_mounting/

4. worked a bit on arch installer script. it's still far from ready but
it's not too bad. the idea's there...

5. wrote a script to pick a line from a file full of quotes so that I
can print one when I start a terminal. it's not very useful but it's not
too bad ^_^

6. updated readme.md as I think I should document my config more
thoroughly in case I ever publish it & if I ever have issues they'll be
documented.

7. found an answer to my issue with df. apparently a service is messing
it up. Anywho, the fix is in this file, I only need to run this command
if it ever comes back.

8. deleted useful_programs and made a more thorough list using pacman
-Qet